Refactor game view into modular components
